[numerics] try to remove codacy warnings

+++ numerics/src/tools/NumericsMatrix.c
@@ -1083,7 +1083,7 @@ void NM_display(const NumericsMatrix* const m)
     printf("========== internalData->iWorkSize = %lu\n", m->internalData->iWorkSize);
     printf("========== internalData->iWork = %p\n", m->internalData->iWork);
 
-    printf("========== internalData->dWorkSize = %lu\n", m->internalData->dWorkSize);
+    printf("========== internalData->dWorkSize = %lu\n", (unsigned long) m->internalData->dWorkSize);
     printf("========== internalData->dWork = %p\n", m->internalData->dWork);
   }
   else
